+++ app/src/main/java/com/rookia/gotflights/domain/network/model/ExchangeRate.kt
@@ -1,8 +1,6 @@
 package com.rookia.gotflights.domain.network.model
 
-import java.math.BigDecimal
-
 data class ExchangeRate(
     val currency: String?,
-    val exchangeRate: BigDecimal?
+    val exchangeRate: Double?
 )
