+++ app/src/main/java/com/rookia/gotflights/domain/network/model/FlightApiResponse.kt
@@ -1,10 +1,9 @@
 package com.rookia.gotflights.domain.network.model
 
-import java.math.BigDecimal
 
 data class FlightApiResponse(
     val currency: String?,
     val inbound: InboundApiResponse?,
     val outbound: OutboundApiResponse?,
-    val price: BigDecimal?
+    val price: Double?
 )
