+++ app/src/test/java/com/rookia/gotflights/utils/MoneyFormatUtilsKtTest.kt
@@ -7,25 +7,25 @@ class MoneyFormatUtilsKtTest {
 
     @Test
     fun `format non zero decimal value with show decimals flag`() {
-        val decimal = 10.5.toBigDecimal()
+        val decimal = 10.5
         assertEquals("10,50", formatDecimalValue(decimal, false))
     }
 
     @Test
     fun `format non zero decimal value with hide decimals flag`() {
-        val decimal = 10.5.toBigDecimal()
+        val decimal = 10.5
         assertEquals("10,5", formatDecimalValue(decimal, true))
     }
 
     @Test
     fun `format zero decimal value with show decimals flag`() {
-        val decimal = 10.0.toBigDecimal()
+        val decimal = 10.0
         assertEquals("10,00", formatDecimalValue(decimal, false))
     }
 
     @Test
     fun `format zero decimal value with hide decimals flag`() {
-        val decimal = 10.0.toBigDecimal()
+        val decimal = 10.0
         assertEquals("10", formatDecimalValue(decimal, true))
     }
 }
