+++ app/src/test/java/com/rookia/gotflights/utils/RangeBarValuesTest.kt
@@ -9,36 +9,36 @@ class RangeBarValuesTest {
 
     @Test
     fun `next value (lower than 10)`() {
-        val value = 8.toBigDecimal()
-        assertEquals(10.toBigInteger(), RangeBarValues.getHigherValue(value))
+        val value = 8.toDouble()
+        assertEquals(10, RangeBarValues.getHigherValue(value))
     }
 
     @Test
     fun `next value (higher than 10)`() {
-        val value = 879.toBigDecimal()
-        assertEquals(880.toBigInteger(), RangeBarValues.getHigherValue(value))
+        val value = 879.toDouble()
+        assertEquals(880, RangeBarValues.getHigherValue(value))
     }
 
     @Test
     fun `prev value (lower than 10)`() {
-        val value = 8.toBigDecimal()
-        assertEquals(0.toBigInteger(), RangeBarValues.getLowerValue(value))
+        val value = 8.toDouble()
+        assertEquals(0, RangeBarValues.getLowerValue(value))
     }
 
     @Test
     fun `prev value (higher than 10)`() {
-        val value = 879.toBigDecimal()
-        assertEquals(870.toBigInteger(), RangeBarValues.getLowerValue(value))
+        val value = 879.toDouble()
+        assertEquals(870, RangeBarValues.getLowerValue(value))
     }
 
     @Test
     fun `get interval `() {
-        assertEquals(2.toBigInteger(), RangeBarValues.getInterval(142.toBigInteger(), 5.toBigInteger()))
+        assertEquals(2, RangeBarValues.getInterval(142, 5))
     }
 
     @Test
     fun `get interval (result less that 1)`() {
-        assertEquals(1.toBigInteger(), RangeBarValues.getInterval(7.toBigInteger(), 5.toBigInteger()))
+        assertEquals(1, RangeBarValues.getInterval(7, 5))
     }
 
     @Test
